05.phrase05/pyproject.toml
[project]
name = "05-phrase05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.22",
    "langchain-openai>=0.3",
    "langgraph>=1.1.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
